+++ tests/test_auth.py
@@ -1,8 +1,11 @@
+import uuid
 from tempfile import NamedTemporaryFile
+from unittest.mock import patch
 
 import pytest
 import yaml
 from chaoslib.exceptions import ActivityFailed
+from yaml.error import YAMLError
 
 from chaosreliably import get_auth_info
 
@@ -24,6 +27,19 @@ def test_using_config_file():
         assert auth_info["token"] == "12345"
         assert auth_info["host"] == "reliably.com"
         assert auth_info["org"] == "test-org"
+
+
+@patch("chaosreliably.yaml.safe_load")
+def test_invalid_yaml_file_errors(mock_safe_load):
+    mock_safe_load.side_effect = YAMLError("An Error")
+    with NamedTemporaryFile(mode="w") as f:
+        f.write("")
+        f.seek(0)
+        with pytest.raises(ActivityFailed) as ex:
+            get_auth_info({"reliably_config_path": f.name})
+    assert str(ex.value) == (
+        f"Failed parsing Reliably configuration at '{f.name}': An Error"
+    )
 
 
 def test_using_config_file_but_override_token_and_host():
@@ -58,30 +74,51 @@ def test_using_secret_only():
 
 
 def test_missing_token_from_secrets():
-    with pytest.raises(ActivityFailed):
+    with pytest.raises(ActivityFailed) as ex:
         get_auth_info(
             {
                 "reliably_config_path": "",
             },
-            {"reliably": {"host": "reliably.dev", "org": "an-org"}},
+            {"host": "reliably.dev", "org": "an-org"},
         )
+    assert str(ex.value) == (
+        "Make sure to provide the Reliably token as a secret or via the Reliably's"
+        " configuration's file."
+    )
 
 
 def test_missing_host_from_secrets():
-    with pytest.raises(ActivityFailed):
+    with pytest.raises(ActivityFailed) as ex:
         get_auth_info(
             {
                 "reliably_config_path": "",
             },
-            {"reliably": {"token": "78890", "org": "an-org"}},
+            {"token": "78890", "org": "an-org"},
         )
+    assert str(ex.value) == (
+        "Make sure to provide the Reliably host as a secret or via the Reliably's "
+        "configuration's file."
+    )
 
 
 def test_missing_org_from_secrets():
-    with pytest.raises(ActivityFailed):
+    with pytest.raises(ActivityFailed) as ex:
         get_auth_info(
             {
                 "reliably_config_path": "",
             },
-            {"reliably": {"token": "78890", "host": "reliably.dev"}},
+            {"token": "78890", "host": "reliably.dev"},
         )
+    assert str(ex.value) == (
+        "Make sure to provide the current Reliably org as a secret or via the"
+        " Reliably's configuration's file."
+    )
+
+
+def test_no_config_at_path_and_no_secrets_provided():
+    with pytest.raises(ActivityFailed) as ex:
+        get_auth_info({"reliably_config_path": str(uuid.uuid4())}, None)
+    assert str(ex.value) == (
+        "Make sure to login against Reliably's services and/or provide the correct"
+        " authentication information to the experiment."
+    )